fix(client): code review comment - move component
