+++ packages/amplication-client/src/Components/CompleteSignupDialog.scss
